compile and run helper
